Fixed problem in MQTT where msgId > 256 is reset to 0

+++ finalmqtt/finalmqtt/Utils/FormatUtil.cs
@@ -29,7 +29,7 @@ namespace mqtttest.Utils
             return mqttString.ToArray();
         }
 
-        public static byte[] toMQttString(int val)
+        public static byte[] toMQttString(short val)
         {
             byte[] data = new byte[2];
             data[0] = (byte)(val >> 8);
